Adds suspense boundary to useSearchParams() for fixing ssr issue
